+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Version History
 v5.27.2
 -------
 
+* Bump d3, d3-geo-veroni, get rid of cypress and vega-lib, update component d3 use of d3.event `<https://github.com/lsst-ts/LOVE-frontend/pull/558>`_
 * Clean compilation warnings on LOVE-frontend `<https://github.com/lsst-ts/LOVE-frontend/pull/557>`_
 
 v5.27.1